support different internal loss datatypes in salrec as well as support for separate val_ndcg

+++ losses/get_loss.py
@@ -2,11 +2,12 @@ from aprec.losses.bpr import BPRLoss
 from aprec.losses.climf import CLIMFLoss
 from aprec.losses.lambdarank import LambdaRankLoss
 from aprec.losses.xendcg import XENDCGLoss
+import tensorflow as tf
 
 
-def get_loss(loss_name, items_num, batch_size, max_positives=40):
+def get_loss(loss_name, items_num, batch_size, max_positives=40, internal_dtype=tf.float32):
     if loss_name == 'lambdarank':
-        return LambdaRankLoss(items_num, batch_size, ndcg_at=max_positives)
+        return LambdaRankLoss(items_num, batch_size, ndcg_at=max_positives, dtype=internal_dtype)
     if loss_name == 'xendcg':
         return XENDCGLoss(items_num, batch_size)
     if loss_name == 'bpr':
